lab2.1_ViewService
lab2_Primary_Backup_Key_Value_Service_1_Viewservice.
we implement a viewservice.The view service goes through a sequence of
numbered views, each with a primary and (if possible) a backup. A view
consists of a view number and the identity (network port name) of the
view's primary and backup servers.