lab3.2_Paxos-based Key/Value Server
In this part,  we build kvpaxos, a fault-tolerant key/value storage
system.  kvpaxos servers will use Paxos to agree on the order in which
client Put()s, Append()s, and Get()s execute. Each time a kvpaxos server
receives a Put()/Append()/Get() RPC, it will use Paxos to cause some
Paxos instance's value to be a description of that operation.( in this
part changed paxos file a lot. according to my classmate, because my
file may not work very well..-_-!)